SDT-8 fixed issue for opening another fragment when clicking multiple times in a trainings fragment.
