+++ app/src/main/java/sdt/tkm/at/steeldarttrainer/training/TrainingsOverViewFragment.kt
@@ -1,7 +1,6 @@
 package sdt.tkm.at.steeldarttrainer.training
 
 import android.app.Fragment
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -127,7 +126,7 @@ class TrainingsOverViewFragment : Fragment() {
 
     private fun replaceFragment(fragment: Fragment) {
         val transaction = fragmentManager.beginTransaction()
-        transaction.add(R.id.content_frame, fragment)
+        transaction.replace(R.id.content_frame, fragment, "Detail_Training")
         transaction.addToBackStack(null)
         transaction.commit()
         oververviewActivity.showUpButton(true)
